move command models as a global model
