+++ prc/models/__init__.py
@@ -7,14 +7,6 @@ Classes to parse and transform PRC API data.
 from .server.status import ServerStatus, ServerOwner, AccountRequirement
 from .server.player import ServerPlayer, QueuedPlayer, PlayerPermission, PlayerTeam
 from .server.vehicle import Vehicle, VehicleName, VehicleModel, VehicleOwner
-from .server.commands import (
-    Command,
-    CommandArg,
-    CommandName,
-    FireType,
-    Weather,
-    CommandTarget,
-)
 from .server.logs import (
     LogEntry,
     LogPlayer,
@@ -25,4 +17,12 @@ from .server.logs import (
     ModCallEntry,
 )
 
-from ..models.player import Player
+from .player import Player
+from .commands import (
+    Command,
+    CommandArg,
+    CommandName,
+    FireType,
+    Weather,
+    CommandTarget,
+)
